Added soft delete for Pathway event cohorts Added codes for Pathway event cohorts Added export for Pathway Analysis Started work on results generation

+++ src/main/java/org/ohdsi/webapi/pathway/repository/PathwayAnalysisEntityRepository.java
@@ -3,5 +3,5 @@ package org.ohdsi.webapi.pathway.repository;
 import com.cosium.spring.data.jpa.entity.graph.repository.EntityGraphJpaRepository;
 import org.ohdsi.webapi.pathway.domain.PathwayAnalysisEntity;
 
-public interface PathwayAnalysisEntityRepository extends EntityGraphJpaRepository<PathwayAnalysisEntity, Long> {
+public interface PathwayAnalysisEntityRepository extends EntityGraphJpaRepository<PathwayAnalysisEntity, Integer> {
 }
